// src/api/growthapiComponents.ts
export interface Habit {
	id: string
	name: string
	description: string
	streak: number
	completed: boolean
	category: string
	userId: string
	createdAt: string
	updatedAt: string
}